essay/requirements.txt
promptflow
azure-storage-blob